Tests/LibWeb/Text/input/HTML/HTMLProgressElement-large-max-value.html

<!DOCTYPE html>
<progress max="1000000000000000000000"></progress>
<script src="../include.js"></script>
<script>
    test(() => {
        const progressElement = document.querySelector("progress");
        println(`progressElement.max: ${progressElement.max}`);
    });
</script>
