Tests/LibWeb/Text/expected/HTML/HTMLProgressElement-large-max-value.txt

progressElement.max: 1e+21
